+++ lib/platform-bible-utils/package.json
@@ -38,7 +38,8 @@
     "lint:scripts": "cross-env NODE_ENV=development eslint --ext .cjs,.js,.jsx,.ts,.tsx --cache .",
     "lint-fix": "npm run lint-fix:scripts",
     "lint-fix:scripts": "prettier --write \"**/*.{ts,tsx,js,jsx,cjs}\" && npm run lint:scripts",
-    "postinstall": "echo platform-bible-utils postinstall"
+    "postinstall": "echo platform-bible-utils postinstall",
+    "test": "jest --silent"
   },
   "devDependencies": {
     "@types/jest": "^29.5.11",
